refactor(api): prepare machine state controller extraction
Move machine-specific state controller configuration and services behind
dedicated machine modules, and introduce shared state controller config in a
new common crate. This reduces direct coupling to the API crate as a first step
toward splitting the machine state controller into a separate crate.
