+++ crates/api/src/state_controller/machine/config/controller.rs
@@ -1,0 +1,114 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use carbide_state_controller_common::config::StateControllerConfig;
+use carbide_utils::config::as_duration;
+use chrono::Duration;
+use duration_str::deserialize_duration_chrono;
+use serde::{Deserialize, Serialize};
+
+/// MachineStateController related config.
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub struct MachineStateControllerConfig {
+    /// Common state controller configs
+    #[serde(default = "StateControllerConfig::default")]
+    pub controller: StateControllerConfig,
+
+    /// How long should we wait before a DPU goes down for sure.
+    #[serde(
+        default = "MachineStateControllerConfig::dpu_wait_time_default",
+        deserialize_with = "deserialize_duration_chrono",
+        serialize_with = "as_duration"
+    )]
+    pub dpu_wait_time: Duration,
+    /// How long to wait for after power down before power on the machine.
+    #[serde(
+        default = "MachineStateControllerConfig::power_down_wait_default",
+        deserialize_with = "deserialize_duration_chrono",
+        serialize_with = "as_duration"
+    )]
+    pub power_down_wait: Duration,
+    /// After how much time, state machine should retrigger reboot if machine does not call back.
+    #[serde(
+        default = "MachineStateControllerConfig::failure_retry_time_default",
+        deserialize_with = "deserialize_duration_chrono",
+        serialize_with = "as_duration"
+    )]
+    pub failure_retry_time: Duration,
+    /// How long to wait for a health report from the DPU before we assume it's down
+    #[serde(
+        default = "MachineStateControllerConfig::dpu_up_threshold_default",
+        deserialize_with = "deserialize_duration_chrono",
+        serialize_with = "as_duration"
+    )]
+    pub dpu_up_threshold: Duration,
+    /// Duration after which a host is considered unhealthy if scout hasn't reported back
+    #[serde(
+        default = "MachineStateControllerConfig::scout_reporting_timeout_default",
+        deserialize_with = "deserialize_duration_chrono",
+        serialize_with = "as_duration"
+    )]
+    pub scout_reporting_timeout: Duration,
+    /// How long to wait for UEFI boot to complete after rebooting a host
+    #[serde(
+        default = "MachineStateControllerConfig::uefi_boot_wait_default",
+        deserialize_with = "deserialize_duration_chrono",
+        serialize_with = "as_duration"
+    )]
+    pub uefi_boot_wait: Duration,
+}
+
+impl MachineStateControllerConfig {
+    pub fn dpu_wait_time_default() -> Duration {
+        Duration::minutes(5)
+    }
+
+    pub fn power_down_wait_default() -> Duration {
+        Duration::minutes(2)
+    }
+
+    pub fn failure_retry_time_default() -> Duration {
+        Duration::minutes(90)
+    }
+
+    pub fn dpu_up_threshold_default() -> Duration {
+        Duration::minutes(5)
+    }
+
+    fn scout_reporting_timeout_default() -> Duration {
+        Duration::minutes(5)
+    }
+
+    pub fn uefi_boot_wait_default() -> Duration {
+        Duration::minutes(5)
+    }
+}
+
+impl Default for MachineStateControllerConfig {
+    fn default() -> Self {
+        Self {
+            controller: StateControllerConfig::default(),
+            dpu_wait_time: MachineStateControllerConfig::dpu_wait_time_default(),
+            power_down_wait: MachineStateControllerConfig::power_down_wait_default(),
+            failure_retry_time: MachineStateControllerConfig::failure_retry_time_default(),
+            dpu_up_threshold: MachineStateControllerConfig::dpu_up_threshold_default(),
+            scout_reporting_timeout: MachineStateControllerConfig::scout_reporting_timeout_default(
+            ),
+            uefi_boot_wait: MachineStateControllerConfig::uefi_boot_wait_default(),
+        }
+    }
+}
