+++ crates/api/src/state_controller/machine/handler.rs
@@ -86,6 +86,9 @@ use model::resource_pool::common::CommonPools;
 use model::site_explorer::ExploredEndpoint;
 use sku::{handle_bom_validation_requested, handle_bom_validation_state};
 use sqlx::PgConnection;
+use state_controller::state_handler::{
+    StateHandler, StateHandlerContext, StateHandlerError, StateHandlerOutcome,
+};
 use tokio::fs::File;
 use tokio::io::{AsyncBufReadExt, AsyncReadExt};
 use tokio::sync::Semaphore;
@@ -93,20 +96,17 @@ use tracing::instrument;
 use version_compare::Cmp;
 
 use crate::CarbideError;
-use crate::cfg::file::{
-    BomValidationConfig, CarbideConfig, MachineValidationConfig, PowerManagerOptions, TimePeriod,
-};
+use crate::cfg::file::{MachineValidationConfig, TimePeriod};
 use crate::dpf::DpfOperations;
 use crate::redfish::{
     self, host_power_control, host_power_control_with_location, set_host_uefi_password,
 };
-use crate::state_controller::common_services::CommonStateHandlerServices;
-use crate::state_controller::machine::context::MachineStateHandlerContextObjects;
+use crate::state_controller::machine::config::{FirmwareGlobal, MachineStateHandlerSiteConfig};
+use crate::state_controller::machine::context::{
+    MachineStateHandlerContextObjects, MachineStateHandlerServices,
+};
 use crate::state_controller::machine::{
     MeasuringOutcome, get_measuring_prerequisites, handle_measuring_state,
-};
-use crate::state_controller::state_handler::{
-    StateHandler, StateHandlerContext, StateHandlerError, StateHandlerOutcome,
 };
 
 mod attestation;
@@ -120,8 +120,9 @@ use helpers::{
     ReprovisionStateHelper, all_equal,
 };
 use rpc::forge_agent_control_response::FileArtifact;
+use state_controller::db_write_batch::DbWriteBatch;
 
-use crate::state_controller::db_write_batch::DbWriteBatch;
+use crate::state_controller::machine::config::{BomValidationConfig, PowerManagerOptions};
 use crate::state_controller::machine::write_ops::MachineWriteOp;
 
 // We can't use http::StatusCode because libredfish has a newer version
@@ -882,7 +883,7 @@ impl MachineStateHandler {
                     let reprov_state = ReprovisionState::next_substate_based_on_bfb_support(
                         self.enable_secure_boot,
                         mh_snapshot,
-                        ctx.services.site_config.dpf.enabled,
+                        ctx.services.site_config.dpf_enabled,
                     );
 
                     let next_state = reprov_state.next_state_with_all_dpus_updated(
@@ -1523,7 +1524,7 @@ impl MachineStateHandler {
                             CleanupState::Init,
                             CleanupContext::Deprovision,
                         );
-                        if !ctx.services.site_config.spdm.enabled {
+                        if !ctx.services.site_config.spdm_enabled {
                             return Ok(StateHandlerOutcome::transition(next_skip_state));
                         }
                         match spdm_measuring_state {
@@ -1561,7 +1562,7 @@ impl MachineStateHandler {
                 spdm_measuring_state,
             } => {
                 let next_skip_state = ManagedHostState::StartAssignmentCycle;
-                if !ctx.services.site_config.spdm.enabled {
+                if !ctx.services.site_config.spdm_enabled {
                     return Ok(StateHandlerOutcome::transition(next_skip_state));
                 }
                 match spdm_measuring_state {
@@ -1621,7 +1622,7 @@ impl MachineStateHandler {
                     instance_state: InstanceState::DpaProvisioning,
                 };
 
-                if !ctx.services.site_config.is_dpa_enabled() {
+                if !ctx.services.site_config.dpa_enabled {
                     // If DPA is not enabled, we don't need to do any DPA provisioning.
                     // So go directly to WaitingForDpaToBeReady state, where we will change
                     // the network status of our DPUs.
@@ -1745,7 +1746,7 @@ impl MachineStateHandler {
         let reprov_state = ReprovisionState::next_substate_based_on_bfb_support(
             self.enable_secure_boot,
             state,
-            ctx.services.site_config.dpf.enabled,
+            ctx.services.site_config.dpf_enabled,
         );
         Ok(Some(reprov_state.next_state_with_all_dpus_updated(
             &state.managed_state,
@@ -1802,7 +1803,7 @@ impl MachineStateHandler {
                     ReprovisionState::next_substate_based_on_bfb_support(
                         self.enable_secure_boot,
                         state,
-                        ctx.services.site_config.dpf.enabled,
+                        ctx.services.site_config.dpf_enabled,
                     )
                     .next_state_with_all_dpus_updated(
                         &state.managed_state,
@@ -3356,7 +3357,7 @@ impl DpuMachineStateHandler {
                     DpuDiscoveringState::next_substate_based_on_bfb_support(
                         self.enable_secure_boot,
                         state,
-                        ctx.services.site_config.dpf.enabled,
+                        ctx.services.site_config.dpf_enabled,
                     );
 
                 tracing::info!(
@@ -5010,7 +5011,7 @@ impl StateHandler for HostMachineStateHandler {
                     let next_skip_state = ManagedHostState::HostInit {
                         machine_state: MachineState::WaitingForDiscovery,
                     };
-                    if !ctx.services.site_config.spdm.enabled {
+                    if !ctx.services.site_config.spdm_enabled {
                         return Ok(StateHandlerOutcome::transition(next_skip_state));
                     }
                     match spdm_measuring_state {
@@ -5923,7 +5924,7 @@ impl StateHandler for InstanceStateHandler {
                         let next_state = ReprovisionState::next_substate_based_on_bfb_support(
                             self.enable_secure_boot,
                             mh_snapshot,
-                            ctx.services.site_config.dpf.enabled,
+                            ctx.services.site_config.dpf_enabled,
                         )
                         .next_state_with_all_dpus_updated(
                             &mh_snapshot.managed_state,
@@ -6026,7 +6027,7 @@ impl StateHandler for InstanceStateHandler {
 
                     // Check each DPA interface associated with the machine to make sure the DPA NIC has updated
                     // its network config (setting VNI to zero in this case).
-                    if ctx.services.site_config.is_dpa_enabled() {
+                    if ctx.services.site_config.dpa_enabled {
                         for dpa_interface in &mh_snapshot.dpa_interface_snapshots {
                             // We're heading back to admin and a DPA still in
                             // Provisioning has nothing to ack -- it never
@@ -6236,7 +6237,7 @@ impl StateHandler for InstanceStateHandler {
                     // DPA state controller re-evaluates with the new host value
                     // (READY -> WaitingForSetVNI, triggering SetVNI).
                     let mut txn = ctx.services.db_pool.begin().await?;
-                    if ctx.services.site_config.is_dpa_enabled() {
+                    if ctx.services.site_config.dpa_enabled {
                         for dpa_interface in &mh_snapshot.dpa_interface_snapshots {
                             db::dpa_interface::try_update_network_config(
                                 &mut txn,
@@ -6257,7 +6258,7 @@ impl StateHandler for InstanceStateHandler {
                     // This involves the DPA State Machine sending SetVNI commands to the NICs, and getting
                     // an ACK. If any of the interfaces has not yet heard back the ACk, we will continue to
                     // be in the current state.
-                    if ctx.services.site_config.is_dpa_enabled() {
+                    if ctx.services.site_config.dpa_enabled {
                         for dpa_interface in &mh_snapshot.dpa_interface_snapshots {
                             if !dpa_interface.managed_host_network_config_version_synced() {
                                 return Ok(StateHandlerOutcome::wait(
@@ -7300,7 +7301,7 @@ impl HostUpgradeState {
         // temporary check if manual upgrade is required before proceeding with automatic ones,
         // should be removed once we complete upgrades through the scout.
         // For now, only gb200s need manual upgrades.
-        if requires_manual_firmware_upgrade(state, &ctx.services.site_config) {
+        if requires_manual_firmware_upgrade(state, &ctx.services.site_config.firmware_global) {
             tracing::info!(
                 "Machine {} (GB200) requires manual firmware upgrade, transitioning to WaitingForManualUpgrade",
                 machine_id
@@ -7734,7 +7735,7 @@ impl HostUpgradeState {
     async fn pre_update_resets(
         &self,
         state: &ManagedHostStateSnapshot,
-        services: &CommonStateHandlerServices,
+        services: &MachineStateHandlerServices,
         scenario: HostFirmwareScenario,
         phase: Option<InitialResetPhase>,
         last_time: &Option<DateTime<Utc>>,
@@ -8696,9 +8697,9 @@ pub async fn host_power_state(
 
 fn requires_manual_firmware_upgrade(
     state: &ManagedHostStateSnapshot,
-    config: &CarbideConfig,
+    firmware_global: &FirmwareGlobal,
 ) -> bool {
-    if !config.firmware_global.requires_manual_upgrade {
+    if !firmware_global.requires_manual_upgrade {
         return false;
     }
 
@@ -9295,7 +9296,7 @@ pub async fn handler_host_power_control_with_location(
 
 async fn restart_dpu(
     machine: &Machine,
-    services: &CommonStateHandlerServices,
+    services: &MachineStateHandlerServices,
     dpf_used_for_ingestion: bool,
 ) -> Result<(), StateHandlerError> {
     let dpu_redfish_client = services.create_redfish_client_from_machine(machine).await?;
@@ -9303,8 +9304,8 @@ async fn restart_dpu(
     // We have seen the boot order be reset on DPUs in some edge cases (for example, after upgrading the BMC and CEC on BF3s)
     // This should take care of handling such cases. It is a no-op most of the time.
     // Skip for DPUs that get their BFB installed via redfish or DPF, they don't need to HTTP boot.
-    let redfish_install = machine.bmc_info.supports_bfb_install()
-        && services.site_config.dpu_config.dpu_enable_secure_boot;
+    let redfish_install =
+        machine.bmc_info.supports_bfb_install() && services.site_config.dpu_enable_secure_boot;
 
     if !redfish_install && !dpf_used_for_ingestion {
         let _ = dpu_redfish_client
@@ -9492,7 +9493,7 @@ async fn call_machine_setup_and_handle_no_dpu_error(
     redfish_client: &dyn Redfish,
     boot_interface_mac: Option<&str>,
     expected_dpu_count: usize,
-    site_config: &CarbideConfig,
+    site_config: &MachineStateHandlerSiteConfig,
 ) -> Result<Option<String>, RedfishError> {
     let setup_result = redfish_client
         .machine_setup(
@@ -9505,7 +9506,7 @@ async fn call_machine_setup_and_handle_no_dpu_error(
     match (
         setup_result,
         expected_dpu_count,
-        site_config.site_explorer.allow_zero_dpu_hosts,
+        site_config.allow_zero_dpu_hosts,
     ) {
         (Err(RedfishError::NoDpu), 0, true) => {
             tracing::info!(
@@ -9522,16 +9523,12 @@ async fn set_boot_order_dpu_first_and_handle_no_dpu_error(
     redfish_client: &dyn Redfish,
     boot_interface_mac: &str,
     expected_dpu_count: usize,
-    site_config: &CarbideConfig,
+    allow_zero_dpu_hosts: bool,
 ) -> Result<Option<String>, RedfishError> {
     let setup_result = redfish_client
         .set_boot_order_dpu_first(boot_interface_mac)
         .await;
-    match (
-        setup_result,
-        expected_dpu_count,
-        site_config.site_explorer.allow_zero_dpu_hosts,
-    ) {
+    match (setup_result, expected_dpu_count, allow_zero_dpu_hosts) {
         (Err(RedfishError::NoDpu), 0, true) => {
             tracing::info!(
                 "redfish set_boot_order_dpu_first failed due to there being no DPUs on the host. This is expected as the host has no DPUs, and we are configured to allow this."
@@ -10440,7 +10437,7 @@ async fn set_host_boot_order(
                 redfish_client,
                 &boot_interface_mac.to_string(),
                 mh_snapshot.host_snapshot.associated_dpu_machine_ids().len(),
-                &ctx.services.site_config,
+                ctx.services.site_config.allow_zero_dpu_hosts,
             )
             .await
             {
@@ -10930,8 +10927,13 @@ mod tests {
             .await
             .unwrap();
 
-        let result =
-            call_machine_setup_and_handle_no_dpu_error(client.as_ref(), None, 1, &config).await;
+        let result = call_machine_setup_and_handle_no_dpu_error(
+            client.as_ref(),
+            None,
+            1,
+            &config.machine_state_handler_site_config(),
+        )
+        .await;
 
         assert!(result.is_ok());
