+++ crates/api/src/tests/power_shelf_state_controller/mod.rs
@@ -21,9 +21,9 @@ use std::time::Duration;
 use db::power_shelf as db_power_shelf;
 use model::power_shelf::PowerShelfControllerState;
 use rpc::forge::forge_server::Forge;
+use state_controller::config::IterationConfig;
 use tokio_util::sync::CancellationToken;
 
-use crate::state_controller::config::IterationConfig;
 use crate::state_controller::controller::StateController;
 use crate::state_controller::power_shelf::context::PowerShelfStateHandlerServices;
 use crate::state_controller::power_shelf::handler::PowerShelfStateHandler;
