+++ crates/api/src/tests/rack_state_controller/mod.rs
@@ -32,9 +32,9 @@ use model::rack::{
 };
 use rpc::forge::StateHistoryRecord;
 use rpc::forge::forge_server::Forge;
+use state_controller::config::IterationConfig;
 use tokio_util::sync::CancellationToken;
 
-use crate::state_controller::config::IterationConfig;
 use crate::state_controller::controller::StateController;
 use crate::state_controller::rack::context::RackStateHandlerContextObjects;
 use crate::state_controller::rack::io::RackStateControllerIO;
