+++ crates/api/src/tests/switch_state_controller/mod.rs
@@ -25,9 +25,9 @@ use db::switch as db_switch;
 use forge_secrets::credentials::TestCredentialManager;
 use model::switch::{ConfiguringState, SwitchControllerState};
 use rpc::forge::forge_server::Forge;
+use state_controller::config::IterationConfig;
 use tokio_util::sync::CancellationToken;
 
-use crate::state_controller::config::IterationConfig;
 use crate::state_controller::controller::StateController;
 use crate::tests::common;
 use crate::tests::common::api_fixtures::create_test_env;
