+++ crates/state-controller-common/src/lib.rs
@@ -15,10 +15,4 @@
  * limitations under the License.
  */
 
-pub mod common_services;
-pub(crate) mod external_service_error;
-pub mod machine;
-pub mod power_shelf;
-pub mod rack;
-
-pub use ::state_controller::{controller, db_write_batch, state_change_emitter, state_handler};
+pub mod config;
